apps/api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
tensorflow-cpu>=2.15.0,<2.17.0
joblib>=1.3.0
yfinance>=0.2.36
python-multipart>=0.0.9
